bugfix: avoid inconsistent use of sample period in DOM.
When multiple functions read the sample period from the DOM, there is a large
window when the sample period has changed, but the 'change' DOM event has not
fired.

This patch stores the sample period in a global variable and changes all the
calculations to use this variable, so that when the global variable changes,
the data series are properly discarded.

Using a global variable is also faster than continuous DOM lookups for every
websocket message received.